requirements.txt
ipykernel
nltk
numpy
pandas
pytypedstream
pyyaml
simplejson
spacy
